Fix CLIP ENVINS actions (issue 6).
For ramps, it now adds a bunch of little steps instead of trying to set two breakpoints, which doesn't seem to be part of the API in live 10 or live 11.  Also fixes param range for PAN envelope, and allows specifying min > max in range to get a negative graph. Finally, switches what used to be a SAW wave to 'TRI' and adds an actual SAW.